backend/requirements.txt
Django==6.0.3
djangorestframework==3.16.1
django-cors-headers==4.9.0
django-filter==25.2
psycopg2-binary==2.9.10
dj-database-url==2.3.0
pillow==12.1.1
python-dotenv==1.2.2
requests==2.32.3
beautifulsoup4==4.13.4
lxml==5.3.1